Add results to db
