+++ app/models/content.rb
@@ -1,7 +1,6 @@
 class Content < ApplicationRecord
   has_many :users, through: :user_contents
 
-  serialize :titles_api_result, JSON
-  serialize :expiring_api_result, JSON
-  serialize :countries_api_result, JSON
+  # serialize :details, JSON
+  # serialize :cast, Array
 end
