+++ app/views/contents/index.html.erb
@@ -1,12 +1,22 @@
 <div class="dashboard-container">
 
+
+  <%= form_tag contents_path, method: :get do %>
+    <%= text_field_tag :query,
+      params[:query],
+      class: "form-control",
+      placeholder: "Find a movie"
+    %>
+    <%= submit_tag "Search", class: "btn btn-primary" %>
+  <% end %>
+
   <% @weeks.keys.each do |weekNumber| %>
     <div class="weekly-content">
       <h3> <%= (Date.commercial(2017, weekNumber).strftime("%A, %b %d"))%> - <%= ((Date.commercial(2017, weekNumber) + 7).strftime("%A, %b %d"))%> </h3>
 
       <div class="cards-container dashboard">
-        <% @weeks[weekNumber].each do |cont| %>
-        <%= render '/shared/content_card', content:cont %>
+        <% @weeks[weekNumber].each do |content| %>
+        <%= render '/shared/content_card', content: content %>
         <% end %>
       </div>
     </div>
