+++ app/views/shared/_content_card.html.erb
@@ -1,8 +1,8 @@
 <div class="card-content dashboard">
-  <%= link_to content_path(content[:imdbid]) do %>
-  <img src="<%= content[:image_url] %>" />
+  <%= link_to content_path(content) do %>
+  <img src="<%= content.image_url %>" />
   <div class="card-content-info">
-    <p> <%= content[:expiration_date] %></p>
+    <p> <%= content.expiration_date %></p>
     <% end %>
   </div>
 </div>
