+++ db/migrate/20200302165450_add_details_to_users.rb
@@ -3,6 +3,5 @@ class AddDetailsToUsers < ActiveRecord::Migration[5.2]
     add_column :users, :first_name, :string
     add_column :users, :last_name, :string
     add_column :users, :country, :string
-    add_column :users, :netflix_details, :string
   end
 end
