+++ db/migrate/20200305151751_add_rating_to_contents.rb
@@ -1,0 +1,5 @@
+class AddRatingToContents < ActiveRecord::Migration[5.2]
+  def change
+    add_column :contents, :netflix_rating, :float
+  end
+end
