+++ db/migrate/20200305155210_add_netflix_id_to_contents.rb
@@ -1,0 +1,5 @@
+class AddNetflixIdToContents < ActiveRecord::Migration[5.2]
+  def change
+    add_index :contents, :netflix_id
+  end
+end
